Cropped Images
This file contains cropped sign images. It was made because I originally thought that the original images file was not registering properly due to the signs being too small in the images in relation to their environment. The signs in these files have been cropped to take up to about 70-90% of the images